Lots of stuff
So I was able to explore all of the datasets and pull out a bunch of info from dr lindgren's stuff that he gave us

I think the most interesting data we have to work with is going to be the parks that charge entrance fees (as it relates to the spec questions). I have to keep working on getting these downloaded as their own csvs so we don't have to process them each time, but so far there is a lot of interesting information. I think i am getting a goof sense of how we should filter down and plot stuff. Also I made a filter function that I added to the utils. it filters for one characteristic, but it can be used for any attribute from the object. I used it here to filter down to NPs only, but I think it will come in handy as we start working with the data!